+++ src/javascript/ZeroClipboard/event.js
@@ -70,6 +70,19 @@ ZeroClipboard.prototype.receiveEvent = function (eventName, args) {
     _removeClass(element, this.options.activeClass);
     break;
 
+  case 'datarequested':
+    var targetId = element.getAttribute('data-clipboard-target'),
+       targetEl = document.getElementById(targetId);
+    if (targetEl) {
+      var textContent = targetEl.value || targetEl.textContent || targetEl.innerText;
+      if (textContent) this.setText(textContent);
+    }
+    else {
+      var defaultText = element.getAttribute('data-clipboard-text');
+      if (defaultText) this.setText(defaultText);
+    }
+    break;
+
   case 'complete':
     this.options.text = null;
     break;